analytics/sales-analytics/requirements.txt
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlalchemy==2.0.20
scikit-learn==1.3.0
plotly==5.16.1
schedule==1.2.0
python-telegram-bot==20.6
loguru==0.7.0 